+++ TvhClient/TVHChannelStoreViewController.m
@@ -235,7 +235,7 @@
     return cell;
 }
 
-- (float)tableView:(UITableView *)tableView heightForFooterInSection:(NSInteger)section {
+- (CGFloat)tableView:(UITableView *)tableView heightForFooterInSection:(NSInteger)section {
     return 0.01f;
 }
